tests: fix logic around sequences
Sequences used by the paper run 01-07 … 91-97, where the first digit
is the price code (0-9) and the second (1-7) is the ISO weekday number.

However, sequences outside this range (00, 08, 09 … 90, 98, 99) *are*
valid and can serve as special codes, so they should not be rejected
if they're specified.

There was also a lingering error where I was happy to pass around
the sequence as an integer (which is fine) but then kicking up a
fuss if it wasn't 2-digits long when converted into a str. Obviously
this causes a problem for valid sequences with a price code of 0,
which are represented by the integers 1-7 (ordinarily) and 0-9 (in
special cases).