star_barcode: fix typo in -dSAFER and, er, disable it
-dSAFER disallows access to other files (good!) but
that catches the bwipp library (bad!), so we've
ended up with -dNOSAFER.

The ghostscript team explain the use of -dSAFER as being
important if you're executing arbitrary PostScript files,
as it denies access to the filesystem. But we're only
executing a script written ourselves and a library that's
both popular and edited down to the minimum requirements
for our script.